Only keep notification as the component
