--- src/Notification/NotificationItem.module.scss
@@ -1,112 +1,0 @@
-@import '../mixins.scss';
-@import '../variables.scss';
-
-.wrapper {
-    @include text-normal;
-
-    border-radius: $border-radius;
-    display: inline-flex;
-    padding: 15px;
-    margin-bottom: 15px;
-    align-items: flex-start;
-}
-
-.default {
-    color: var(--grey-dark-2);
-    background-color: var(--white);
-    fill: var(--black);
-
-    :global(.night-mode) & {
-        color: var(--black);
-    }
-
-    .title {
-        color: var(--black);
-    }
-
-    .closeIcon {
-        fill: var(--grey-dark);
-    }
-}
-
-.icon {
-    margin: 5px 0;
-}
-
-.closeIcon {
-    cursor: pointer;
-    width: 13px;
-    height: auto;
-}
-
-.titleIcon {
-    width: 22px;
-    height: auto;
-    margin: 2px 15px 0 0;
-}
-
-.title {
-    font-weight: bold;
-    margin-bottom: 6px;
-}
-
-.messageArea {
-    width: 100%;
-}
-
-.info {
-    fill: var(--blue);
-}
-
-.warn {
-    fill: var(--orange);
-}
-
-.error {
-    fill: var(--red);
-}
-
-.success {
-    fill: var(--green);
-}
-
-.solid {
-    color: var(--white);
-    fill: var(--white);
-
-    .actionButton {
-        color: var(--white);
-    }
-}
-
-.infoSolid {
-    background-color: var(--blue);
-}
-
-.warnSolid {
-    background-color: var(--orange);
-}
-
-.errorSolid {
-    background-color: var(--red);
-}
-
-.successSolid {
-    background-color: var(--green);
-}
-
-.actionButtonArea {
-    display: flex;
-    margin-top: 10px;
-    align-items: center;
-}
-
-.actionButton {
-    color: var(--green);
-    cursor: pointer;
-    margin-right: 20px;
-
-    &:last-child {
-        margin-right: 0;
-    }
-}
